A kind whose bundles are catalog rows refuses bytes at every door
An MCP server's document lives in the server catalog, so there is nothing for a
publish or a proposal to carry. Both doors say so before any custody call, in the
same sentence, and name the page where the act belongs — rather than ingesting
bytes for a bundle that will never serve them.

What went with it: the server-document gate the byte path ran, the embedded-name
claim it made afterwards, and the advisory probe that followed a publish. The
final transaction can no longer refuse anything, so it is an ordinary transaction
again — and the suite that proved a claim never outlives its move now proves what
is left, which is that a genesis publish and its receipt are one write.
